requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
openpyxl==3.1.2
tenacity==8.2.3
aiohttp==3.9.3
pytest==8.0.0
pytest-cov==4.1.0 